Added support for multiple model loading
